:recycle: 責務毎にメソッドを分割
